+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.3] - 2024-04-17
+
+- Fixed MQTT reconnection bug
+- Fixed resend all states every 24h
+- Updated all examples to use the new connect method
+
 ## [1.0.2] - 2024-03-18
 
 - Added *unit_of_measurement* configuration to HASensorNumeric